src/main/java/domain/service/LadderGame.java
package domain.service;

import domain.model.Ladder;
import domain.model.Players;

public class LadderGame {
    private final Ladder ladder;
    private final Players players;

    public LadderGame(Ladder ladder, Players players) {
        this.ladder = ladder;
        this.players = players;
    }

    public void play() {
        for (int i = 0; i < ladder.getHeight(); i++) {
            players.moveAll(ladder);
        }
    }

    public Players getPlayers() {
        return players;
    }
}
